// frontend/src/models/student.ts
export interface Student {
  studentNumber: string //ID
  lastName: string
  firstName: string
  fullName: string
  email: string
  personalEmail: string
  arrivalGroup: string
  administrativeGroup: string
  degreeProgramme: string
  formOfEducation: string
  enrollment: string
  assessment: string
}